+++ contracts/periphery/interfaces/IUSUMRouter.sol
@@ -9,7 +9,7 @@ interface IUSUMRouter is IUSUMLiquidityCallback {
         address settlementToken,
         uint256 takerMargin,
         uint256 makerMargin,
-        int256 qty,
+        int224 qty,
         uint32 leverage,
         uint256 deadline
     ) external;
